Add input mocking end-to-end test
Keyboard input can now be mocked in tests, by building
the test with mock_keyboard_input.cpp instead of the
actual keyboard_input.cpp.
This allows tests to queue up mocked user input, to
simulate an end-to-end test.

The program loop is extracted outside the main application's
entrypoint file, which allows it to be called from tests.
To verify results during the program loop, the stdout
is captured and evaluated.
