+++ src/functions.h
@@ -3,8 +3,6 @@
 #include "voting_round.h"
 
 /* -------------- Command line inputs -------------- */
-auto getKey() -> char;
-auto getConfirmation() -> bool;
 auto continueWithoutSaving(std::optional<VotingRound> const& voting_round, std::string const& str) -> bool;
 
 /* -------------- Menu alternatives -------------- */
