+++ src/keyboard_input.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <string>
+
+auto getKey() -> char;
+auto getConfirmation() -> bool;
+auto getLine() -> std::string;
